+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added *Fill Caps* option to *Curve Object Output* node.
 - Added *Bevel Mode* option to *Curve Object Output* node.
 - Added *Paramater* option to *Spline Falloff* node.
+- Added *Wrap Paramaters* option to *Evaluate Spline* node.
 
 ### Fixed